Allow using RSA model as base for DirectRefGame
via a pretty ugly hack that overrides the --listener flag.
Basically, RSA can either be a speaker or a listener, and
DirectRefGame can accept either as its base model (but always
produces a listener model). So --listener must always be True for
the correct evaluation data to be passed to DirectRefGame, but
then we need to explicitly specify whether RSA should be passed
--listener=True or --listener=False.